Refactor useAuth hook to normal functions There is no need for state or effects with auth, so it does not need to be in a hook
