--- src/app/components/hooks/use-auth.spec.tsx
@@ -1,4 +1,0 @@
-// TODO: Implement this
-describe('useAuth', () => {
-  it('should work', () => {});
-});
